Refactor memory

+++ obligatory_assignment/include/binary_file.h
@@ -4,8 +4,7 @@
 #include <stdio.h>  // for FILE
 struct Router;
 
-int readBinaryFile(const char* binFile, struct Router** routerArray,
-                   unsigned int* N);
+int readBinaryFile(const char* binFile, struct Router** routerArray, size_t* N);
 int readRouter(FILE* fp, struct Router* router);
 
 #endif  // OBLIGATORY_ASSIGNMENT_INCLUDE_BINARY_FILE_H_
